requirements.txt
alembic==1.13.1
bcrypt
bleach
blinker==1.7.0
boltons 
Brotli 
certifi 
cffi 
charset-normalizer 
click==8.1.7
cryptography 
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
Flask==3.0.2
Flask-Babel
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Restful
Flask-Security-Too
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna 
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch 
jsonpointer==2.1
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient 
packaging 
pluggy 
pycosat 
pycparser 
pyOpenSSL 
PySocks
python-dotenv
requests 
ruamel.yaml 
ruamel.yaml.clib 
setuptools
SQLAlchemy==2.0.29
toolz 
tqdm 
typing_extensions==4.10.0
urllib3 
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
zstandard 
